// frontend/src/types/generated/security-lgpd/conectaPROLGPDSecurityAPI.schemas.ts
export type StandardResponseData = { [key: string]: unknown };

export interface StandardResponse {
  /** Indica se a operação foi bem-sucedida */
  success: boolean;
  /** Mensagem descritiva do resultado */
  message: string;
  /** Dados retornados pela operação */
  data?: StandardResponseData;
}
// ...
export type ErasureRequestSchemaScope = typeof ErasureRequestSchemaScope[keyof typeof ErasureRequestSchemaScope];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ErasureRequestSchemaScope = {
  all: 'all',
  personal: 'personal',
  transactional: 'transactional',
} as const;

export interface ErasureRequestSchema {
  /** UUID do titular */
  titular_id: string;
  /** Email do titular */
  titular_email: string;
  /** Motivo da exclusão */
  reason: string;
  /** Escopo da exclusão */
  scope?: ErasureRequestSchemaScope;
}